+++ app/models/time_offset.rb
@@ -1,0 +1,4 @@
+# Represents time offsets for users
+class TimeOffset < ApplicationRecord
+  belongs_to :user
+end