+++ app/serializers/time_offset_serializer.rb
@@ -1,0 +1,3 @@
+class TimeOffsetSerializer < ActiveModel::Serializer
+  attributes :id, :value
+end