+++ db/migrate/20191112211155_create_time_offsets.rb
@@ -1,0 +1,10 @@
+class CreateTimeOffsets < ActiveRecord::Migration[5.2]
+  def change
+    create_table :time_offsets do |t|
+      t.integer :user_id
+      t.integer :value
+      t.timestamps
+    end
+    add_index :time_offsets, :user_id
+  end
+end