+++ frontend/packs/application.js
@@ -20,6 +20,7 @@ import "components/craft/craft";
 import "components/static_management/static_management";
 import "components/guild_management/guild_management";
 import "components/guild_invite_new/guild_invite_new";
+import "components/settings/settings";
 
 import Rails from "rails-ujs";