+++ spec/factories/time_offsets.rb
@@ -1,0 +1,6 @@
+FactoryBot.define do
+  factory :time_offset do
+    value { nil }
+    association :user
+  end
+end