+++ spec/models/user_spec.rb
@@ -3,6 +3,7 @@ RSpec.describe User, type: :model do
   it { should have_many(:guilds).through(:characters) }
   it { should have_many(:subscribes).through(:characters) }
   it { should have_many(:identities).dependent(:destroy) }
+  it { should have_one(:time_offset).dependent(:destroy) }
   it { should validate_presence_of :email }
   it { should validate_presence_of :password }
   it { should validate_presence_of :role }