+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.5] - 2019-11-13
+### Added
+- settings page
+- TimeOffset model for user preferences about rendered time
+- selecting prefered time for users
+- render all times for user time zone
+
 ## [0.7.4] - 2019-11-12
 ### Added
 - api endpoint recipes#search